+++ client/src/pages/game/[displayPosition]/index.page.tsx
@@ -13,6 +13,7 @@ import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react';
 import { Image, Layer, Stage, Text } from 'react-konva';
 import Boom from 'src/components/Effect/Boom';
+import { FreedomMeteor } from 'src/components/Effect/FreedomMeteor';
 import { Meteor } from 'src/components/Effect/Meteor';
 import { Bullet } from 'src/components/Entity/Bullet';
 import { Enemy } from 'src/components/Entity/Enemy';
@@ -126,6 +127,9 @@ const Game = () => {
         </Layer>
         <Layer>
           <Meteor displayPosition={displayPosition ?? 0} />
+        </Layer>
+        <Layer>
+          <FreedomMeteor />
         </Layer>
         <Layer>
           {bullets.map((bullet) => (
